requirements.txt
glfw==2.5.5
numpy==1.23.4
PyOpenGL==3.1.6
